requirements.txt
httpx==0.27.2
pyodbc==5.1.0
psycopg[binary]==3.3.3
python-dotenv==1.0.1